# apps/runner/pkg/api/docs/swagger.yaml
definitions:
...
  CreateBackupDTO:
    properties:
      registry:
        $ref: '#/definitions/RegistryDTO'
      snapshot:
        type: string
    required:
    - registry
    - snapshot
    type: object
  CreateSandboxDTO:
    properties:
      authToken:
        type: string
      cpuQuota:
        minimum: 1
        type: integer
      domainAllowList:
        type: string
      entrypoint:
        items:
          type: string
        type: array
      env:
        additionalProperties:
          type: string
        type: object
      fromVolumeId:
        type: string
      gpuQuota:
        minimum: 0
        type: integer
      id:
        type: string
      linkedSandboxId:
        description: |-
          LinkedSandboxId identifies an existing sandbox this sandbox should be co-located with.
          When set, the runner should attach both sandboxes to a shared local network so they can communicate.
        type: string
      memoryQuota:
        minimum: 1
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        description: |-
          Name is the human-readable sandbox name. When present, it is used as the
          network alias for link networks so linked sandboxes can resolve each other
          by name.
        type: string
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      organizationId:
        description: Nullable for backward compatibility
        type: string
      osUser:
        type: string
      otelEndpoint:
        type: string
      regionId:
        type: string
      registry:
        $ref: '#/definitions/RegistryDTO'
      sandboxClass:
        description: Optional for backward compatibility, but when provided, indicates
          the class of sandbox to create.
        type: string
      skipStart:
        type: boolean
      snapshot:
        type: string
      storageQuota:
        minimum: 1
        type: integer
      userId:
        type: string
      volumeMountCredentials:
        $ref: '#/definitions/dto.VolumeMountCredentialsDTO'
      volumes:
        items:
          $ref: '#/definitions/dto.VolumeDTO'
        type: array
    required:
    - id
    - osUser
    - snapshot
    - userId
    type: object
...
  ErrorResponse:
    description: Error response
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    required:
    - message
    - path
    - statusCode
    - timestamp
    type: object
...
  IsRecoverableDTO:
    properties:
      errorReason:
        type: string
    required:
    - errorReason
    type: object
  IsRecoverableResponse:
    properties:
      recoverable:
        type: boolean
    type: object
...
  RecoverSandboxDTO:
    properties:
      backupErrorReason:
        type: string
      cpuQuota:
        minimum: 1
        type: integer
      env:
        additionalProperties:
          type: string
        type: object
      errorReason:
        description: At least one of ErrorReason or BackupErrorReason must yield a
          recovery type; both are optional.
        type: string
      fromVolumeId:
        type: string
      gpuQuota:
        minimum: 0
        type: integer
      memoryQuota:
        minimum: 1
        type: integer
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      osUser:
        type: string
      registry:
        $ref: '#/definitions/RegistryDTO'
      snapshot:
        type: string
      storageQuota:
        minimum: 1
        type: integer
      userId:
        type: string
      volumeMountCredentials:
        $ref: '#/definitions/dto.VolumeMountCredentialsDTO'
      volumes:
        items:
          $ref: '#/definitions/dto.VolumeDTO'
        type: array
    required:
    - osUser
    - userId
    type: object
...
  SandboxInfoResponse:
    properties:
      backupError:
        type: string
      backupSnapshot:
        type: string
      backupState:
        $ref: '#/definitions/enums.BackupState'
      daemonVersion:
        type: string
      recoverable:
        type: boolean
      state:
        $ref: '#/definitions/enums.SandboxState'
    type: object
...
  StartSandboxResponse:
    properties:
      daemonVersion:
        type: string
    type: object
  StopSandboxDTO:
    properties:
      force:
        type: boolean
    type: object
...
  UpdateNetworkSettingsDTO:
    properties:
      domainAllowList:
        type: string
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      networkLimitEgress:
        type: boolean
    type: object
...
  dto.VolumeMountCredentialsDTO:
    additionalProperties:
      $ref: '#/definitions/dto.VolumeMountCredentialDTO'
    type: object
...
info:
  contact: {}
  description: Daytona Runner API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona Runner API
  version: v0.0.0-dev
paths:
...
  /sandboxes:
    post:
      description: Create a sandbox
      operationId: Create
      parameters:
      - description: Create sandbox
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/CreateSandboxDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/StartSandboxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}:
    get:
      description: Get sandbox info
      operationId: Info
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox info
          schema:
            $ref: '#/definitions/SandboxInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get sandbox info
      tags:
      - sandbox
  /sandboxes/{sandboxId}/backup:
    post:
      description: Create sandbox backup
      operationId: CreateBackup
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Create backup
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/CreateBackupDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Backup started
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create sandbox backup
      tags:
      - sandbox
  /sandboxes/{sandboxId}/destroy:
    post:
      description: Destroy sandbox
      operationId: Destroy
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox destroyed
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Destroy sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/is-recoverable:
    post:
      consumes:
      - application/json
      description: Check if the sandbox's error reason indicates a recoverable error
      operationId: IsRecoverable
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Error reason to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/IsRecoverableDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/IsRecoverableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Check if sandbox error is recoverable
      tags:
      - sandbox
  /sandboxes/{sandboxId}/network-settings:
    get:
      description: Get sandbox network settings
      operationId: GetNetworkSettings
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Network settings
          schema:
            $ref: '#/definitions/UpdateNetworkSettingsDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get sandbox network settings
      tags:
      - sandbox
    post:
      description: Update sandbox network settings
      operationId: UpdateNetworkSettings
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Update network settings
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/UpdateNetworkSettingsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Network settings updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update sandbox network settings
      tags:
      - sandbox
  /sandboxes/{sandboxId}/recover:
    post:
      consumes:
      - application/json
      description: Recover sandbox from error state using specified recovery type
      operationId: Recover
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Recovery parameters
        in: body
        name: recovery
        required: true
        schema:
          $ref: '#/definitions/RecoverSandboxDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox recovered
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Recover sandbox from error state
      tags:
      - sandbox
  /sandboxes/{sandboxId}/resize:
    post:
      description: Resize sandbox
      operationId: Resize
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Resize sandbox
        in: body
        name: sandbox
        required: true
        schema:
          $ref: '#/definitions/ResizeSandboxDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox resized
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Resize sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/snapshot-from-sandbox:
    post:
      description: Commit the sandbox container filesystem and push the image to the
        supplied registry under the canonical daytona-{hash}:daytona tag.
      operationId: SnapshotFromSandbox
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Snapshot from sandbox
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateSnapshotFromSandboxRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SnapshotInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Snapshot a running sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/start:
    post:
      description: Start sandbox
      operationId: Start
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Start options
        in: body
        name: sandbox
        schema:
          $ref: '#/definitions/StartSandboxDTO'
      - description: Auth token
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox started
          schema:
            $ref: '#/definitions/StartSandboxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Start sandbox
      tags:
      - sandbox
  /sandboxes/{sandboxId}/stop:
    post:
      description: Stop sandbox
      operationId: Stop
      parameters:
      - description: Sandbox ID
        in: path
        name: sandboxId
        required: true
        type: string
      - description: Stop sandbox
        in: body
        name: sandbox
        schema:
          $ref: '#/definitions/StopSandboxDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sandbox stopped
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Stop sandbox
      tags:
      - sandbox
...
security:
- Bearer: []
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and an API token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"